Added user-specified delimiter as text field (does not quite work yet). Moved some duplicated code to the networkTask, just before it calls the string query. Set some default values.

+++ src/main/java/dk/ku/cpr/proteoVisualizer/internal/ui/SearchOptionsPanel.java
@@ -66,7 +66,7 @@ public class SearchOptionsPanel extends JPanel {
 	public SearchOptionsPanel(final AppManager manager) {
 		super(new GridBagLayout());
 		this.manager = manager;
-		confidence = (int)(manager.getDefaultConfidence());
+		confidence = manager.getDefaultConfidence();
 		this.networkType = manager.getDefaultNetworkType();
 		initOptions();
 		setPreferredSize(new Dimension(700,125));
